Convert table reservation form to a modal
Add a modal that allows the users to reserve a schedule. The modal will
be accessed by pressing a button in the jumbotron. The modal contains
the number of guests, the section (smoking or non-smoking) and the
schedule. The user can also directly call the restaurant by pressing a
link inside an alert in the modal.